split up unit and integration tests
